refactor: changes after code review
