+++ src/features/auth/ui/signUp/signUp.tsx
@@ -2,7 +2,7 @@ import { SentEmail } from '@/features'
 import { GithubSvgrepoCom31, GoogleSvgrepoCom1 } from '@/shared/assets'
 import { ROUTES } from '@/shared/config'
 import { Trans } from '@/shared/ui'
-import { useTranslation } from '@/shared/utils/hooks/useTranslation'
+import { useTranslation } from '@/shared/utils'
 import { Button, Card, FormCheckbox, FormInput, Typography } from '@photo-fiesta/ui-lib'
 import Link from 'next/link'
