+++ src/shared/config/i18n/locales/en.ts
