+++ src/shared/config/i18n/locales/index.ts
@@ -1,0 +1,2 @@
+export { en } from './en'
+export { ru } from './ru'
