+++ src/shared/config/i18n/locales/ru.ts
