+++ src/shared/utils/hooks/useTranslation.ts
@@ -1,6 +1,5 @@
-import { en } from 'locales/en'
-import { ru } from 'locales/ru'
 import { useRouter } from 'next/router'
+import { en, ru } from 'src/shared/config/i18n/locales'
 
 export const useTranslation = () => {
   const router = useRouter()
